+++ src/components/common/Post/Post.tsx
@@ -9,21 +9,12 @@ import PostData from "@/components/page-layout/HomeLayout/types";
 import { ROUTE } from "@/constants/route";
 import Heart from "@/icons/heart.svg";
 import RedHeart from "@/icons/red_haert.svg";
+import { formatDateString } from "@/utils";
 
 const cn = classNames.bind(styles);
 
 interface PostProps {
   data: PostData;
-}
-
-function formatDateString(dateString: string) {
-  const date = new Date(dateString);
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-
-  return `${year}.${month}.${day}`;
 }
 
 export default function Post({ data }: PostProps) {
